requirements.txt
appdirs==1.4.4
asgiref==3.8.1
blessed==1.21.0
certifi==2025.8.3
charset-normalizer==3.4.3
curtsies==0.4.3
cwcwidth==0.1.10
Django==5.2.5
django-debug-toolbar==6.0.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-nested-admin==4.1.1
django-nested-inline==0.4.6
django-revproxy==0.13.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
jedi==0.19.2
Markdown==3.8.2
packaging==25.0
parsedatetime==2.6
parso==0.8.4
pep8==1.7.1
prompt_toolkit==3.0.51
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
pyxdg==0.28
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
wcwidth==0.2.13
